+++ equalsverifier-release-nodep/src/main/java/module-info.java
@@ -1,0 +1,13 @@
+module nl.jqno.equalsverifier {
+    exports nl.jqno.equalsverifier;
+    exports nl.jqno.equalsverifier.api;
+
+    // Built-in prefab values
+    requires static com.google.common;
+    requires static java.desktop;
+    requires static java.naming;
+    requires static java.rmi;
+    requires static java.sql;
+    requires static javafx.base;
+    requires static org.joda.time;
+}