+++ equalsverifier-release-verify/src/test/java/nl/jqno/equalsverifier/verify_release/jar/MainClassesJarTest.java
@@ -1,5 +1,7 @@
 package nl.jqno.equalsverifier.verify_release.jar;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import nl.jqno.equalsverifier.verify_release.jar.helper.JarAsserter;
 import nl.jqno.equalsverifier.verify_release.jar.helper.JarReader;
 import org.junit.jupiter.api.AfterAll;
@@ -46,5 +48,19 @@ class MainClassesJarTest {
     @Test
     void versionsOfClassFiles() {
         jar.assertVersionsOfClassFiles();
+    }
+
+    @Test
+    void presenceOfModuleInfoWithDependencies() {
+        jar.assertModuleInfoWithDependencies();
+    }
+
+    @Test
+    void moduleInfoIsIdenticalToCore() throws Exception {
+        try (var coreReader = new JarReader("equalsverifier-core.jar")) {
+            var coreModuleinfo = coreReader.getContentOf("module-info.class");
+            var mainModuleinfo = reader.getContentOf("module-info.class");
+            assertThat(mainModuleinfo).containsExactly(coreModuleinfo);
+        }
     }
 }